strands_agents/requirements.txt
# ============================================
# Strands Agents & AWS Bedrock AgentCore
# ============================================
strands-agents>=0.1.0            # Strands Agents framework
boto3>=1.34.0                     # AWS SDK for Python
botocore>=1.34.0                  # Low-level AWS service client
anthropic>=0.18.0                 # Claude LLM client

# ============================================
# AWS Bedrock Runtime & Agent Services
# ============================================
boto3-stubs[bedrock-runtime]>=1.34.0
boto3-stubs[bedrock-agent-runtime]>=1.34.0
boto3-stubs[cognito-idp]>=1.34.0
boto3-stubs[secretsmanager]>=1.34.0

# ============================================
# Neo4j Graph Database
# ============================================
neo4j>=5.14.0                     # Official Neo4j Python driver
neo4j-driver>=5.14.0              # Neo4j driver with advanced features

# ============================================
# Web Framework & API (Development Proxy)
# ============================================
fastapi>=0.109.0                  # Modern async web framework
uvicorn[standard]>=0.27.0         # ASGI server
pydantic>=2.5.0                   # Data validation
pydantic-settings>=2.1.0          # Settings management

# ============================================
# Authentication & Security
# ============================================
python-jose[cryptography]>=3.3.0  # JWT token handling
cryptography>=42.0.0              # Encryption utilities

# ============================================
# Utilities & Configuration
# ============================================
python-dotenv>=1.0.0              # Environment variable management
pyyaml>=6.0.1                     # YAML configuration
click>=8.1.0                      # CLI tool creation

# ============================================
# Logging & Monitoring
# ============================================
structlog>=24.1.0                 # Structured logging
python-json-logger>=2.0.7         # JSON log formatting

# ============================================
# Async & Performance
# ============================================
aiohttp>=3.9.0                    # Async HTTP client
aioboto3>=12.3.0                  # Async AWS SDK

# ============================================
# Development & Testing
# ============================================
pytest>=7.4.0                     # Testing framework
pytest-asyncio>=0.21.0            # Async test support
pytest-cov>=4.1.0                 # Code coverage
httpx>=0.26.0                     # Async HTTP testing client
moto[all]>=5.0.0                  # Mock AWS services
